write the exitstatus in an atomic operation
Sometimes we already tried to read from the status file although
we haven't yet written the exitstatus to the file, because we
might have been blocked by other operations or the timing was
just too perfect. We we then barfed on an empty string as exitcode.
This commit ensures that if the file is present, there will also
be an exit status written to it.